[M144-LTS] Vulkan: Cap memory allocation size to 1GB
  Currently the maximum memory allocation size for a single object
can be inquired from the Vulkan driver via KHR_maintenance3 (promoted
to core in Vulkan 1.1). However, the reported limit can still be quite
large for common usage. In addition, it can increase the risk of memory
size calculations overflowing if the operands are not 64 bits. As an
example, if the limit is 4GB and an image is defined with 32-bit dims
and format that theoretically require 4GB, the multiplication of these
values can result in overflow, even if it is to be assigned to a 64-bit
value.
  * (One way to avoid such overflow cases is to split multiplication
    into several steps, e.g., CL:7595734)

This change aims to reduce the maximum size allowed for allocation in
order to reduce the risk of such overflow issues.

* Added kMemoryAllocationSizeLimit to vk_renderer.
  * Currently set to 1GB.

* Added to vk::Renderer: mMaxMemoryAllocationSize
  * Set to the minimum of kMemoryAllocationSizeLimit and
    the maxMemoryAllocationSize reported from the driver.

* Cast the values used in some buffer size calculations to
  size_t.
  * (The leftmost value being size_t should be enough to propagate
    this type for the result of each multiplication.)

* Updated the following test to use a smaller width and height to
  avoid error due to the new cap:
  TextureCubeTestES32.MaxArrayTextureLayersVerify

Bug: chromium:486972661
Bug: chromium:487208468
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/7615650
Commit-Queue: Amirali Abdolrashidi <abdolrashidi@google.com>
(cherry picked from commit 4de47461e45248eeaf8fd0ef04ca3949f98029da)
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/7691581
Commit-Queue: Gyuyoung Kim (xWF) <qkim@google.com>

+++ src/libANGLE/renderer/vulkan/vk_helpers.cpp
@@ -10488,9 +10488,10 @@ angle::Result ImageHelper::copyImageDataToBuffer(ContextVk *contextVk,
     // used in this function to be of some combined depth and stencil format.
     ASSERT(getAspectFlags() == VK_IMAGE_ASPECT_COLOR_BIT);
 
-    uint32_t pixelBytes = imageFormat.pixelBytes;
-    size_t bufferSize =
-        sourceArea.width * sourceArea.height * sourceArea.depth * pixelBytes * layerCount;
+    size_t pixelBytes = imageFormat.pixelBytes;
+    size_t bufferSize = static_cast<size_t>(sourceArea.width) *
+                        static_cast<size_t>(sourceArea.height) *
+                        static_cast<size_t>(sourceArea.depth) * pixelBytes * layerCount;
 
     const VkImageAspectFlags aspectFlags = getAspectFlags();
